+++ runtime/v2/builder.go
@@ -19,24 +19,24 @@ import (
 // AppBuilder is a type that is injected into a container by the runtime/v2 module
 // (as *AppBuilder) which can be used to create an app which is compatible with
 // the existing app.go initialization conventions.
-type AppBuilder struct {
-	app          *App
+type AppBuilder[T transaction.Tx] struct {
+	app          *App[T]
 	storeOptions *rootstore.FactoryOptions
 
 	// the following fields are used to overwrite the default
 	branch      func(state store.ReaderMap) store.WriterMap
-	txValidator func(ctx context.Context, tx transaction.Tx) error
-	postTxExec  func(ctx context.Context, tx transaction.Tx, success bool) error
+	txValidator func(ctx context.Context, tx T) error
+	postTxExec  func(ctx context.Context, tx T, success bool) error
 }
 
 // DefaultGenesis returns a default genesis from the registered AppModule's.
-func (a *AppBuilder) DefaultGenesis() map[string]json.RawMessage {
+func (a *AppBuilder[T]) DefaultGenesis() map[string]json.RawMessage {
 	return a.app.moduleManager.DefaultGenesis()
 }
 
 // RegisterModules registers the provided modules with the module manager.
 // This is the primary hook for integrating with modules which are not registered using the app config.
-func (a *AppBuilder) RegisterModules(modules ...appmodulev2.AppModule) error {
+func (a *AppBuilder[T]) RegisterModules(modules ...appmodulev2.AppModule) error {
 	for _, appModule := range modules {
 		if mod, ok := appModule.(appmodule.HasName); ok {
 			name := mod.Name()
@@ -62,7 +62,7 @@ func (a *AppBuilder) RegisterModules(modules ...appmodulev2.AppModule) error {
 // This method should only be used for registering extra stores
 // wiich is necessary for modules that not registered using the app config.
 // To be used in combination of RegisterModules.
-func (a *AppBuilder) RegisterStores(keys ...string) {
+func (a *AppBuilder[T]) RegisterStores(keys ...string) {
 	a.app.storeKeys = append(a.app.storeKeys, keys...)
 	if a.storeOptions != nil {
 		a.storeOptions.StoreKeys = append(a.storeOptions.StoreKeys, keys...)
@@ -70,7 +70,7 @@ func (a *AppBuilder) RegisterStores(keys ...string) {
 }
 
 // Build builds an *App instance.
-func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
+func (a *AppBuilder[T]) Build(opts ...AppBuilderOption[T]) (*App[T], error) {
 	for _, opt := range opts {
 		opt(a)
 	}
@@ -87,7 +87,7 @@ func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
 
 	// default post tx exec
 	if a.postTxExec == nil {
-		a.postTxExec = func(ctx context.Context, tx transaction.Tx, success bool) error {
+		a.postTxExec = func(ctx context.Context, tx T, success bool) error {
 			return nil
 		}
 	}
@@ -96,21 +96,12 @@ func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
 		return nil, err
 	}
 
-	stfMsgHandler, err := a.app.msgRouterBuilder.Build()
-	if err != nil {
-		return nil, fmt.Errorf("failed to build STF message handler: %w", err)
-	}
-
-	stfQueryHandler, err := a.app.queryRouterBuilder.Build()
-	if err != nil {
-		return nil, fmt.Errorf("failed to build query handler: %w", err)
-	}
-
 	endBlocker, valUpdate := a.app.moduleManager.EndBlock()
 
-	a.app.stf = stf.NewSTF[transaction.Tx](
-		stfMsgHandler,
-		stfQueryHandler,
+	stf, err := stf.NewSTF[T](
+		a.app.logger.With("module", "stf"),
+		a.app.msgRouterBuilder,
+		a.app.queryRouterBuilder,
 		a.app.moduleManager.PreBlocker(),
 		a.app.moduleManager.BeginBlock(),
 		endBlocker,
@@ -119,6 +110,11 @@ func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
 		a.postTxExec,
 		a.branch,
 	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create STF: %w", err)
+	}
+
+	a.app.stf = stf
 
 	rs, err := rootstore.CreateRootStore(a.storeOptions)
 	if err != nil {
@@ -126,7 +122,7 @@ func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
 	}
 	a.app.db = rs
 
-	appManagerBuilder := appmanager.Builder[transaction.Tx]{
+	appManagerBuilder := appmanager.Builder[T]{
 		STF:                a.app.stf,
 		DB:                 a.app.db,
 		ValidateTxGasLimit: a.app.config.GasConfig.ValidateTxGasLimit,
@@ -159,33 +155,33 @@ func (a *AppBuilder) Build(opts ...AppBuilderOption) (*App, error) {
 }
 
 // AppBuilderOption is a function that can be passed to AppBuilder.Build to customize the resulting app.
-type AppBuilderOption func(*AppBuilder)
+type AppBuilderOption[T transaction.Tx] func(*AppBuilder[T])
 
 // AppBuilderWithBranch sets a custom branch implementation for the app.
-func AppBuilderWithBranch(branch func(state store.ReaderMap) store.WriterMap) AppBuilderOption {
-	return func(a *AppBuilder) {
+func AppBuilderWithBranch[T transaction.Tx](branch func(state store.ReaderMap) store.WriterMap) AppBuilderOption[T] {
+	return func(a *AppBuilder[T]) {
 		a.branch = branch
 	}
 }
 
 // AppBuilderWithTxValidator sets the tx validator for the app.
 // It overrides all default tx validators defined by modules.
-func AppBuilderWithTxValidator(txValidators func(ctx context.Context, tx transaction.Tx) error) AppBuilderOption {
-	return func(a *AppBuilder) {
+func AppBuilderWithTxValidator[T transaction.Tx](txValidators func(ctx context.Context, tx T) error) AppBuilderOption[T] {
+	return func(a *AppBuilder[T]) {
 		a.txValidator = txValidators
 	}
 }
 
 // AppBuilderWithPostTxExec sets logic that will be executed after each transaction.
 // When not provided, a no-op function will be used.
-func AppBuilderWithPostTxExec(
+func AppBuilderWithPostTxExec[T transaction.Tx](
 	postTxExec func(
 		ctx context.Context,
-		tx transaction.Tx,
+		tx T,
 		success bool,
 	) error,
-) AppBuilderOption {
-	return func(a *AppBuilder) {
+) AppBuilderOption[T] {
+	return func(a *AppBuilder[T]) {
 		a.postTxExec = postTxExec
 	}
 }