+++ schema/appdata/listener.go
@@ -1,7 +1,9 @@
-package indexerbase
+package appdata
 
 import (
 	"encoding/json"
+
+	"cosmossdk.io/schema"
 )
 
 // Listener is an interface that defines methods for listening to both raw and logical blockchain data.
@@ -31,6 +33,7 @@ type Listener struct {
 	OnEvent func(EventData) error
 
 	// OnKVPair is called when a key-value has been written to the store for a given module.
+	// Module names must conform to the NameFormat regular expression.
 	OnKVPair func(moduleName string, key, value []byte, delete bool) error
 
 	// Commit is called when state is committed, usually at the end of a block. Any
@@ -43,13 +46,13 @@ type Listener struct {
 	// should ensure that they have performed whatever initialization steps (such as database
 	// migrations) required to receive OnObjectUpdate events for the given module. If the
 	// indexer's schema is incompatible with the module's on-chain schema, the listener should return
-	// an error.
-	InitializeModuleSchema func(module string, schema ModuleSchema) error
+	// an error. Module names must conform to the NameFormat regular expression.
+	InitializeModuleSchema func(moduleName string, moduleSchema schema.ModuleSchema) error
 
 	// OnObjectUpdate is called whenever an object is updated in a module's state. This is only called
 	// when logical data is available. It should be assumed that the same data in raw form
-	// is also passed to OnKVPair.
-	OnObjectUpdate func(module string, update ObjectUpdate) error
+	// is also passed to OnKVPair. Module names must conform to the NameFormat regular expression.
+	OnObjectUpdate func(moduleName string, update schema.ObjectUpdate) error
 }
 
 // InitializationData represents initialization data that is passed to a listener.