+++ server/v2/commands.go
@@ -7,15 +7,38 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"strings"
 	"syscall"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 
 	"cosmossdk.io/core/transaction"
 	"cosmossdk.io/log"
 )
 
-func Commands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logger log.Logger, components ...ServerComponent[transaction.Tx]) (CLIConfig, error) {
+// Execute executes the root command of an application.
+// It handles adding core CLI flags, specifically the logging flags.
+func Execute(rootCmd *cobra.Command, envPrefix, defaultHome string) error {
+	rootCmd.PersistentFlags().String(FlagLogLevel, "info", "The logging level (trace|debug|info|warn|error|fatal|panic|disabled or '*:<level>,<key>:<level>')")
+	rootCmd.PersistentFlags().String(FlagLogFormat, "plain", "The logging format (json|plain)")
+	rootCmd.PersistentFlags().Bool(FlagLogNoColor, false, "Disable colored logs")
+	rootCmd.PersistentFlags().StringP(FlagHome, "", defaultHome, "directory for config and data")
+
+	// update the global viper with the root command's configuration
+	viper.SetEnvPrefix(envPrefix)
+	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_", "-", "_"))
+	viper.AutomaticEnv()
+
+	return rootCmd.Execute()
+}
+
+func Commands[AppT AppI[T], T transaction.Tx](
+	rootCmd *cobra.Command,
+	newApp AppCreator[AppT, T],
+	logger log.Logger,
+	components ...ServerComponent[AppT, T],
+) (CLIConfig, error) {
 	if len(components) == 0 {
 		return CLIConfig{}, errors.New("no components provided")
 	}
@@ -42,7 +65,7 @@ func Commands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logger 
 
 			app := newApp(l, v)
 
-			if _, err := server.Init(app, v, l); err != nil {
+			if err := server.Init(app, v, l); err != nil {
 				return err
 			}
 
@@ -76,7 +99,12 @@ func Commands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logger 
 	return cmds, nil
 }
 
-func AddCommands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logger log.Logger, components ...ServerComponent[transaction.Tx]) error {
+func AddCommands[AppT AppI[T], T transaction.Tx](
+	rootCmd *cobra.Command,
+	newApp AppCreator[AppT, T],
+	logger log.Logger,
+	components ...ServerComponent[AppT, T],
+) error {
 	cmds, err := Commands(rootCmd, newApp, logger, components...)
 	if err != nil {
 		return err
@@ -90,8 +118,7 @@ func AddCommands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logg
 			return err
 		}
 
-		err = configHandle(server, home, cmd)
-		if err != nil {
+		if err = configHandle(server, home, cmd); err != nil {
 			return err
 		}
 
@@ -108,14 +135,17 @@ func AddCommands(rootCmd *cobra.Command, newApp AppCreator[transaction.Tx], logg
 }
 
 // configHandle writes the default config to the home directory if it does not exist and sets the server context
-func configHandle(s *Server, home string, cmd *cobra.Command) error {
-	if _, err := os.Stat(filepath.Join(home, "config")); os.IsNotExist(err) {
-		if err = s.WriteConfig(filepath.Join(home, "config")); err != nil {
+func configHandle[AppT AppI[T], T transaction.Tx](s *Server[AppT, T], home string, cmd *cobra.Command) error {
+	configDir := filepath.Join(home, "config")
+
+	// we need to check app.toml as the config folder can already exist for the client.toml
+	if _, err := os.Stat(filepath.Join(configDir, "app.toml")); os.IsNotExist(err) {
+		if err = s.WriteConfig(configDir); err != nil {
 			return err
 		}
 	}
 
-	viper, err := ReadConfig(filepath.Join(home, "config"))
+	viper, err := ReadConfig(configDir)
 	if err != nil {
 		return err
 	}